+++ domaindriven-examples/postgres/Event/V1.hs
@@ -1,5 +1,6 @@
 module Event.V1 where
 
+import Control.DeepSeq (NFData)
 import Data.Aeson (FromJSON, ToJSON)
 import GHC.Generics (Generic)
 import Prelude
@@ -7,4 +8,4 @@ import Prelude
 data CounterEvent
     = CounterIncreased
     | CounterDecreased
-    deriving (Show, Generic, FromJSON, ToJSON)
+    deriving (Show, Generic, FromJSON, ToJSON, NFData)