+++ domaindriven-examples/simple/Main.hs
@@ -8,6 +8,7 @@
 {-# LANGUAGE OverloadedRecordDot #-}
 module Main where
 
+import Control.DeepSeq (NFData)
 import Control.Monad (when)
 import Data.Aeson (FromJSON, ToJSON)
 import Data.Time (UTCTime)
@@ -39,7 +40,7 @@ data CounterModel = CounterModel
 data CounterEvent
     = CounterIncreased
     | CounterDecreased
-    deriving (Show, Generic, ToJSON, FromJSON)
+    deriving (Show, Generic, ToJSON, FromJSON, NFData)
 
 --------------------------------------------------------------------------------
 -- Define event handler